:recycle: Move paperwallet to utils folder
